requirements.txt
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
Pillow==11.3.0
numpy==2.2.6
scikit-learn==1.7.1
joblib==1.5.1
requests==2.32.4
python-dotenv==1.1.1
boto3==1.35.82